+++ spring-telqos-core/src/main/java/com/dwarfeng/springtelqos/node/configuration/SpringTelqosQosDefinitionParser.java
@@ -2,7 +2,6 @@ package com.dwarfeng.springtelqos.node.configuration;
 
 import com.dwarfeng.springtelqos.impl.service.TelqosQosServiceImpl;
 import com.dwarfeng.springtelqos.sdk.util.BeanDefinitionParserUtil;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.config.ConstructorArgumentValues;
 import org.springframework.beans.factory.config.RuntimeBeanReference;
@@ -24,11 +23,11 @@ public class SpringTelqosQosDefinitionParser implements BeanDefinitionParser {
 
     @Override
     public BeanDefinition parse(Element element, @Nonnull ParserContext parserContext) {
-        String handlerName = BeanDefinitionParserUtil.mayResolvePlaceholder(
-                parserContext, element.getAttribute("handler-name")
-        );
         String serviceName = BeanDefinitionParserUtil.mayResolvePlaceholder(
                 parserContext, element.getAttribute("service-name")
+        );
+        String handlerRef = BeanDefinitionParserUtil.mayResolvePlaceholder(
+                parserContext, element.getAttribute("handler-ref")
         );
         String semRef = BeanDefinitionParserUtil.mayResolvePlaceholder(
                 parserContext, element.getAttribute("sem-ref")
@@ -37,19 +36,6 @@ public class SpringTelqosQosDefinitionParser implements BeanDefinitionParser {
                 parserContext, element.getAttribute("auto-start")
         );
 
-        if (StringUtils.isEmpty(serviceName)) {
-            serviceName = "telqosQosService";
-        }
-        if (StringUtils.isEmpty(handlerName)) {
-            handlerName = "telqosHandlerImpl";
-        }
-        if (StringUtils.isEmpty(semRef)) {
-            semRef = "mapServiceExceptionMapper";
-        }
-        if (StringUtils.isEmpty(autoStart)) {
-            autoStart = "true";
-        }
-
         BeanDefinitionParserUtil.makeSureBeanNameNotDuplicated(parserContext, serviceName);
 
         BeanDefinitionBuilder telqosQosServiceBuilder = BeanDefinitionBuilder.rootBeanDefinition(
@@ -57,7 +43,7 @@ public class SpringTelqosQosDefinitionParser implements BeanDefinitionParser {
         );
         telqosQosServiceBuilder.getRawBeanDefinition().setAutowireMode(AbstractBeanDefinition.AUTOWIRE_CONSTRUCTOR);
         ConstructorArgumentValues telqosQosServiceConstructorArgumentValues = new ConstructorArgumentValues();
-        telqosQosServiceConstructorArgumentValues.addIndexedArgumentValue(0, new RuntimeBeanReference(handlerName));
+        telqosQosServiceConstructorArgumentValues.addIndexedArgumentValue(0, new RuntimeBeanReference(handlerRef));
         telqosQosServiceConstructorArgumentValues.addIndexedArgumentValue(1, new RuntimeBeanReference(semRef));
         telqosQosServiceBuilder.getRawBeanDefinition().setConstructorArgumentValues(
                 telqosQosServiceConstructorArgumentValues